04-My-First-Rag-System/requirements_groq.txt
python-dotenv
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq
chromadb
huggingface-hub
sentence-transformers